refactor: este commit irá alterar os nomes dos arquivos de types e corrigir os imports
