+++ src/templates/Base/index.tsx
@@ -1,6 +1,6 @@
 import { Container, Header } from "@/components";
 
-import { BaseTemplateProps } from "./base";
+import { BaseTemplateProps } from "./types";
 
 const Base = ({ children }: BaseTemplateProps) => {
   return (
